# Avoid `Expr` copies `OptimizeProjection`, 12% faster planning, encapsulate indicies